+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  - Add /wand restore command for re-creating a bound wand
  - Fix bypass_confusion parameter
  - Improve "magic damage", now with entity source
+ - Familiars don't drop loot when killed
  
 ## 3.5